Push all requests to the queue before crawling
Thanks @kmike for the catch, that's very important!
The reason why not all requests from start_requests were sent
to the queue was that we started downloading from the first
seeds, and after a while the downloader became always busy
with requests gathered from the first seeds, never having a
chance to push other seeds into the queue.